Enemy Spawer Sync: New Enemy and recv EnemyId from Server

+++ TAKOEngine/Rendering/SpriteDX12.h
@@ -100,7 +100,7 @@ public:
 	int GetTextureHeight() const { return m_texture_height; }
 
 	const DirectX::XMFLOAT2 GetTextureSize() const {
-		return DirectX::XMFLOAT2(m_texture_width, m_texture_height);
+		return DirectX::XMFLOAT2((float)(m_texture_width), (float)(m_texture_height));
 	}
 
 	// ディスクリプタ取得
